+++ CHANGELOG.md
@@ -19,7 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `cli`: Ignore errors from `enable_vector_catch` (#1714).
 - `cli`: Retry RTT attach before continuing (#1722).
 - `cli`: Clean clap attributes (#xxxx)
-- `target-gen` : Memory regions in target.yaml are now sorted with lowest address first (# )
+- `target-gen`: (#1745)
+  - Memory regions in target.yaml are now sorted with lowest address first.
+  - Use `.pdsc` flash algorithm `RAMstart` field to calculate `load_address` for target yaml.
 
 ## [0.20.0]